Added Dark Theme To Telemetry Interface
Added a functionality so the user can click a checkbox on the options
modal to enable a dark theme. The interface reverts to its original
appearance when the checkbox is unchecked. A new css file was added to
have the page use its styling to create a dark theme. A cookie is also stored to remember the user's layout choice, and load the darktheme by default if
checked.