Added Server Autoconnect Functionality
This functionality allows the user to connect a device to Telemetry, and
have its COM port remembered. Later on, if this device is connected again, it should seamlessly connect to the previous TTY device that it was connected it when it 'first' saw it. This is done using local storage in HTML5.