Fixing xterm key events
Before this fix, if you typed '.' or the various escape characters, they
would not be transmitted to the server, because they do not evaluate to
a proper uri.

To fix this, the write route has become:
/write/<cr>/<nl>?payload=<insert characters here>

The payload can be anything as long as it is URI encoded. This includes
the arrow keys, control escape codes like CTRL+C etc.

Added support for home and end keys.